+++ CHANGES.md
@@ -21,6 +21,7 @@
 - Remove custom `@slow` decorator for Pytest marker with `--include-slow` flag
 - Add `get_queues` to `RQ` main class, replacing `Queue.all()`
 - Replace Worker's `find_by_key` with `load_by_key`
+- Replace Worker's `all_keys` with `get_workers_keys` on `RQ` main class
 
 ### RQ 1.15.1 (2023-06-20)